Fully integrated policies to GUI - Encountered a problem in DeleteConditionalDiscount in DB must fix

+++ Version-1/eCommerce/eCommerce/src/DataAccessLayer/DBUtil.cs
@@ -1998,7 +1998,11 @@ namespace eCommerce.src.DataAccessLayer
 
         public void DeleteConditionalDiscount(FilterDefinition<BsonDocument> filter)
         {
+
+                
             DTO_ConditionalDiscount deletedConditionalDiscount = DAO_ConditionalDiscount.Delete(filter);
+            if (deletedConditionalDiscount == null) // fix here deletedConditionDiscount can't be null
+                return;
             foreach (var condition in deletedConditionalDiscount.Condition)
             {
                 DeleteIDiscountCondition(condition.Value, condition.Key);
@@ -2008,6 +2012,7 @@ namespace eCommerce.src.DataAccessLayer
                 DeleteIDiscountPolicy(discount.Value, discount.Key);
             }
             Discount_ConditionalDiscounts.TryRemove(deletedConditionalDiscount._id, out ConditionalDiscount c);
+
         }
